Support for different RPM sensor types
