+++ EDCmain/RPMCustomCPS.h
@@ -1,6 +1,7 @@
-#ifdef RPM_SENSOR_TYPE_DEFAULT
-#ifndef RPMDEFAULTCPS_H
-#define RPMDEFAULTCPS_H
+#ifdef RPM_SENSOR_TYPE_CUSTOM
+
+#ifndef RPMCUSTOMCPS_H
+#define RPMCUSTOMCPS_H
 
 #include "RPMBase.h"
 #include "Arduino.h"
@@ -14,21 +15,26 @@
 // ((((16000000/64)/65535)*60/5                         16000000L             4000 ~ 780rpm
 #define RPMTIMER_DURATION_TO_RPM(x) ((unsigned long)(60*F_CPU/64/NUMBER_OF_CYLINDERS)/((unsigned long)x))  // 250 000hz frequency (max. 65535 ticks per teeth ~45rpm)
 #define RPMTIMER_MIN_DURATON 400 // 7500rpm
-static volatile unsigned int injectionBegin;
-static volatile unsigned int rawValues[NUMBER_OF_CYLINDERS];
 
-class RPMDefaultCPS : public RPMBase {
+#define RPM_TEETH_COUNT (144)
+#define RPM_TEETH_PER_CYL (144/NUMBER_OF_CYLINDERS)
+
+static volatile unsigned int rawValues[RPM_TEETH_PER_CYL];
+
+
+class RPMCustomCPS : public RPMBase {
 	private:
-	unsigned int storedValues[NUMBER_OF_CYLINDERS];
+	unsigned int storedValues[RPM_TEETH_PER_CYL];
 	public:
 	void init();
 	unsigned int getLatestMeasure();
 	unsigned int getLatestRawValue();
-	int getInjectionTiming();
+	
 	private:
 	void setupTimers();
 
 };
+
 
 #endif
 #endif
